refactor: Move WalletUpdate to wallet module
